pyproject.toml
[tool.poetry]
name = "csu-notice"
version = "0.1.0-alpha.9"
description = "CSU Notice"
authors = ["Jigsaw <j1g5aw@foxmail.com>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0b1"
nonebot-adapter-onebot = "^2.0.0b1"
nonebot-plugin-apscheduler = "^0.1.2"
sqlmodel = "^0.0.6"
pyquery = "^1.4.3"
aiosqlite = "^0.17.0"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.35"}

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"
black = {version = "^21.9b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
